# internal/docs/generated/swagger.yaml
basePath: /api
definitions:
  apicommon.APIErrorBody:
    properties:
      code:
        description: 错误码
        type: string
      details:
        type: object
      message:
        description: 错误描述
        type: string
    type: object
  apicommon.APIErrorResponse:
    properties:
      error:
        allOf:
        - $ref: '#/definitions/apicommon.APIErrorBody'
        description: 错误体
    type: object
...
  v2.V2EnvironmentalAPIResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/v2.V2EnvironmentalResponse'
        description: 响应数据
      meta:
        allOf:
        - $ref: '#/definitions/v2.V2Meta'
        description: 响应元信息
    type: object
  v2.V2EnvironmentalResponse:
    properties:
      summary:
        $ref: '#/definitions/v2.V2EnvironmentalSummary'
      timeseries:
        allOf:
        - $ref: '#/definitions/v2.V2EnvironmentalTimeseries'
        description: 时序
    type: object
  v2.V2EnvironmentalSummary:
    properties:
      battery_heater_minutes:
        description: 电池加热器时长 (分)
        type: number
      climate_on_minutes:
        description: 空调开启时长 (分)
        type: number
      defroster_minutes:
        description: 除霜时长 (分)
        type: number
      elevation_gain_total:
        description: 累计爬升 (米)
        type: number
      elevation_loss_total:
        description: 累计下降 (米)
        type: number
      elevation_max:
        description: 最高海拔 (米)
        type: number
      elevation_min:
        description: 最低海拔 (米)
        type: number
      inside_temp_avg:
        description: 车内平均温 (°C)
        type: number
      outside_temp_avg:
        description: 车外平均温 (°C)
        type: number
      outside_temp_max:
        description: 车外最高温 (°C)
        type: number
      outside_temp_min:
        description: 车外最低温 (°C)
        type: number
    type: object
  v2.V2EnvironmentalTimeseries:
    properties:
      group_by:
        description: 聚合粒度
        type: string
      items:
        description: 条目列表
        items:
          $ref: '#/definitions/v2.V2EnvironmentalTimeseriesItem'
        type: array
    type: object
  v2.V2EnvironmentalTimeseriesItem:
    properties:
      climate_on_ratio:
        type: number
      elevation_avg:
        type: number
      elevation_gain:
        description: 该周期累计爬升 (米)
        type: number
      elevation_loss:
        description: 该周期累计下降 (米)
        type: number
      elevation_max:
        type: number
      elevation_min:
        type: number
      inside_temp_avg:
        description: 车内平均温 (°C)
        type: number
      inside_temp_max:
        description: 车内最高温 (°C)
        type: number
      inside_temp_min:
        description: 车内最低温 (°C)
        type: number
      outside_temp_avg:
        description: 车外平均温 (°C)
        type: number
      outside_temp_max:
        description: 车外最高温 (°C)
        type: number
      outside_temp_min:
        description: 车外最低温 (°C)
        type: number
      period_start:
        description: 周期起始
        type: string
    type: object
...
  v2.V2Meta:
    properties:
      applied_filters:
        additionalProperties:
          type: string
        description: |-
          AppliedFilters echoes back the query-string filters the server actually
          honored (audit §5.2 / spec §5.2). Helps clients debug "why did I get this
          payload?" — they can compare what they sent vs. what the server applied
          (defaults, normalized casing, ignored unknown keys, etc.). Always
          non-nil when populated; serialized as a JSON object.
        type: object
      car_id:
        description: 车辆 ID
        example: 1
        type: integer
      compare:
        description: 对比模式
        enum:
        - none
        - previous_period
        type: string
      end:
        description: 结束时间
        format: date-time
        type: string
      generated_at:
        format: date-time
        type: string
      period:
        description: 聚合周期
        enum:
        - day
        - week
        - month
        - quarter
        - year
        - custom
        - lifetime
        type: string
      start:
        description: 起始时间
        format: date-time
        type: string
      timezone:
        description: 时区
        example: Asia/Shanghai
        type: string
      unit:
        allOf:
        - $ref: '#/definitions/v2.V2Unit'
        description: 单位
    type: object
...
  v2.V2Unit:
    properties:
      consumption:
        example: Wh/km
        type: string
      currency:
        example: CNY
        type: string
      distance:
        description: 距离 (km)
        example: km
        type: string
      duration:
        description: 时长 (秒)
        example: seconds
        type: string
      elevation:
        example: m
        type: string
      energy:
        example: kWh
        type: string
      power:
        example: kW
        type: string
      speed:
        example: km/h
        type: string
      temperature:
        example: C
        type: string
    type: object
...
info:
  contact: {}
  description: TeslaMate 数据查询与分析 API。提供 V1（原始数据）和 V2（聚合分析）两套接口。
  title: TeslaMateApi
  version: "2.0"
paths:
...
  /v2/cars/{CarID}/analytics/environmental:
    get:
      description: 返回该车的车外/车内温度聚合、HVAC 活动时长、海拔统计。
      parameters:
      - description: 车辆 ID
        example: 1
        in: path
        name: CarID
        required: true
        type: integer
      - description: 聚合周期
        enum:
        - day
        - week
        - month
        - quarter
        - year
        - custom
        in: query
        name: period
        type: string
      - description: 起始时间（RFC3339）
        in: query
        name: start
        type: string
      - description: 结束时间（RFC3339）
        in: query
        name: end
        type: string
      - description: IANA 时区
        in: query
        name: timezone
        type: string
      - description: 逗号分隔的扩展项，如 timeseries
        example: '"timeseries"'
        in: query
        name: include
        type: string
      - description: 时序聚合粒度（仅在 include=timeseries 时生效）
        enum:
        - day
        - week
        - month
        - year
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.V2EnvironmentalAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apicommon.APIErrorResponse'
      summary: V2 环境分析（温度/HVAC/海拔）
      tags:
      - v2
...
swagger: "2.0"
...